+++ app/src/main/java/com/example/app/AuthReceiver.kt
@@ -1,0 +1,12 @@
+package com.example.app
+
+import android.content.BroadcastReceiver
+import android.content.Context
+import android.content.Intent
+
+class AuthReceiver(private val onAuthReceived: (String?) -> Unit) : BroadcastReceiver() {
+    override fun onReceive(context: Context?, intent: Intent?) {
+        val authCode = intent?.getStringExtra("authorizationCode")
+        onAuthReceived(authCode)
+    }
+}
